# ENH:  Vectorize np.partition and np.argpartition using AVX-512