--- a/numpy/core/include/numpy/ndarraytypes.h
+++ b/numpy/core/include/numpy/ndarraytypes.h
@@ -404,7 +404,7 @@ typedef int (PyArray_PartitionFunc)(void *, npy_intp, npy_intp,
                                     npy_intp *, npy_intp *, npy_intp,
                                     void *);
 typedef int (PyArray_ArgPartitionFunc)(void *, npy_intp *, npy_intp, npy_intp,
-                                       npy_intp *, npy_intp *,
+                                       npy_intp *, npy_intp *, npy_intp,
                                        void *);
 
 typedef int (PyArray_FillWithScalarFunc)(void *, npy_intp, void *, void *);