--- a/numpy/core/src/npysort/selection.cpp
+++ b/numpy/core/src/npysort/selection.cpp
@@ -36,10 +36,17 @@ inline bool quickselect_dispatch(T* v, npy_intp num, npy_intp kth)
     /*
      * Only defined for int16_t, uint16_t, float16, int32_t, uint32_t, float32,
      * int64_t, uint64_t, double
+     * TODO: Enable quickselect for 32-bit. np.argpartition is only vectorized
+     * for 64-bit when npy_intp is 8 bytes, which means np.partition and
+     * np.argpartition will produces different results on 32-bit systems.
+     * Several tests in test_multiarray.py rely on these results being
+     * identical. We should get rid of this constraint and re-write
+     * these tests once we enable this for 32-bit.
      */
     if constexpr (
         (std::is_integral_v<T> || std::is_floating_point_v<T> || std::is_same_v<T, np::Half>) &&
-        (sizeof(T) == sizeof(uint16_t) || sizeof(T) == sizeof(uint32_t) || sizeof(T) == sizeof(uint64_t))) {
+        (sizeof(T) == sizeof(uint16_t) || sizeof(T) == sizeof(uint32_t) || sizeof(T) == sizeof(uint64_t)) &&
+        sizeof(npy_intp) == sizeof(int64_t)) {
         using TF = typename np::meta::FixedWidth<T>::Type;
         void (*dispfunc)(TF*, npy_intp, npy_intp) = nullptr;
         if constexpr (sizeof(T) == sizeof(uint16_t)) {