--- a/numpy/core/src/npysort/simd_qsort.dispatch.cpp
+++ b/numpy/core/src/npysort/simd_qsort.dispatch.cpp
@@ -16,6 +16,30 @@
 namespace np { namespace qsort_simd {
 
 #if defined(NPY_HAVE_AVX512_SKX) && !defined(_MSC_VER)
+template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(int32_t *arr, npy_intp* arg, npy_intp num, npy_intp kth)
+{
+    avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+}
+template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(uint32_t *arr, npy_intp* arg, npy_intp num, npy_intp kth)
+{
+    avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+}
+template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(int64_t*arr, npy_intp* arg, npy_intp num, npy_intp kth)
+{
+    avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+}
+template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(uint64_t*arr, npy_intp* arg, npy_intp num, npy_intp kth)
+{
+    avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+}
+template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(float *arr, npy_intp* arg, npy_intp num, npy_intp kth)
+{
+    avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+}
+template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(double *arr, npy_intp* arg, npy_intp num, npy_intp kth)
+{
+    avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+}
 template<> void NPY_CPU_DISPATCH_CURFX(QSelect)(int32_t *arr, npy_intp num, npy_intp kth)
 {
     avx512_qselect(arr, kth, num, true);