… benchmarks/benchmarks/bench_function_base.py
… numpy/core/src/npysort/simd_qsort.hpp
… numpy/core/src/npysort/simd_qsort_16bit.dispatch.cpp
… numpy/core/tests/test_multiarray.py