Provides a speed up of **25x** for 16-bit, **17x** for 32-bit dtypes and about **8x** speed up for 64-bit dtypes. Benchmark numbers on an array of size 100000 for varying values of k (10, 100 and 1000). 

Benchmark for random data on Intel TGL: 

```
       before           after         ratio
     [7d18effb]       [0c8fed1f]
     <main>           <np-partition>
-        1.15±0ms          193±1μs     0.17  bench_function_base.Partition.time_partition('float64', ('random',), 10)
-        1.15±0ms        192±0.9μs     0.17  bench_function_base.Partition.time_partition('float64', ('random',), 1000)
-     1.15±0.01ms          192±2μs     0.17  bench_function_base.Partition.time_partition('float64', ('random',), 100)
-        1.13±0ms          135±2μs     0.12  bench_function_base.Partition.time_partition('float32', ('random',), 100)
-        1.13±0ms          135±2μs     0.12  bench_function_base.Partition.time_partition('float32', ('random',), 10)
-         965±1μs          116±2μs     0.12  bench_function_base.Partition.time_partition('int64', ('random',), 100)
-     1.13±0.01ms          135±2μs     0.12  bench_function_base.Partition.time_partition('float32', ('random',), 1000)
-         970±2μs          115±1μs     0.12  bench_function_base.Partition.time_partition('int64', ('random',), 1000)
-        973±10μs        114±0.9μs     0.12  bench_function_base.Partition.time_partition('int64', ('random',), 10)
-         898±2μs       52.2±0.9μs     0.06  bench_function_base.Partition.time_partition('uint32', ('random',), 100)
-         902±6μs       52.2±0.9μs     0.06  bench_function_base.Partition.time_partition('int32', ('random',), 100)
-         904±3μs         52.3±2μs     0.06  bench_function_base.Partition.time_partition('int32', ('random',), 1000)
-         902±6μs         51.8±1μs     0.06  bench_function_base.Partition.time_partition('uint32', ('random',), 10)
-         900±4μs       51.6±0.9μs     0.06  bench_function_base.Partition.time_partition('int32', ('random',), 10)
-         906±4μs         51.9±2μs     0.06  bench_function_base.Partition.time_partition('uint32', ('random',), 1000)
-        1.02±0ms       43.5±0.2μs     0.04  bench_function_base.Partition.time_partition('float16', ('random',), 100)
-        1.02±0ms       43.3±0.2μs     0.04  bench_function_base.Partition.time_partition('float16', ('random',), 10)
-        1.04±0ms       43.4±0.2μs     0.04  bench_function_base.Partition.time_partition('float16', ('random',), 1000)
-        1.14±0ms      45.1±0.09μs     0.04  bench_function_base.Partition.time_partition('int16', ('random',), 1000)
-        1.14±0ms       45.2±0.1μs     0.04  bench_function_base.Partition.time_partition('int16', ('random',), 100)
-        1.14±0ms       45.1±0.3μs     0.04  bench_function_base.Partition.time_partition('int16', ('random',), 10)

```
Benchmark numbers on other kinds of array can be seen [here](https://gist.github.com/r-devulap/d09d05098039fb91ef71dee272deaf38).