@r-devulap 2023-07-17
Looks like all the CI failures are related to 2 tests: 

1. `numpy/lib/tests/test_shape_base.py::test_argequivalent.py`
2.  `numpy/core/tests/test_multiarray.py::test_partition.py` 

where both the tests expect the output of `np.partition(arr, k)` to match `arr[np.argpartition(arr, k)]`. Output of np.partition and np.argpartition ae not unique and hence I assume this isn't a requirement, right?