@r-devulap 2023-07-24
I guess that problem solved itself when we vectorize both partition and argpartition :)