@r-devulap 2023-07-24
Benchmarks for np.argpartition on a TGL laptop: About a **6x** speed up for 32-bit and 64-bit dtypes.

```
       before           after         ratio
     [7d18b532]       [88696f5a]
     <main>           <np-partition>
-     1.10±0.01ms          208±1μs     0.19  bench_function_base.Partition.time_argpartition('int32', ('random',), 100)
-     1.11±0.01ms          207±1μs     0.19  bench_function_base.Partition.time_argpartition('int32', ('random',), 1000)
-     1.11±0.01ms          207±1μs     0.19  bench_function_base.Partition.time_argpartition('int32', ('random',), 10)
-        1.16±0ms          209±2μs     0.18  bench_function_base.Partition.time_argpartition('int64', ('random',), 100)
-        1.16±0ms          207±1μs     0.18  bench_function_base.Partition.time_argpartition('int64', ('random',), 10)
-        1.17±0ms          206±1μs     0.18  bench_function_base.Partition.time_argpartition('int64', ('random',), 1000)
-        1.34±0ms        224±0.9μs     0.17  bench_function_base.Partition.time_argpartition('float64', ('random',), 100)
-        1.34±0ms          223±2μs     0.17  bench_function_base.Partition.time_argpartition('float64', ('random',), 10)
-     1.35±0.01ms        223±0.9μs     0.16  bench_function_base.Partition.time_argpartition('float64', ('random',), 1000)
-     1.34±0.01ms          208±1μs     0.16  bench_function_base.Partition.time_argpartition('float32', ('random',), 100)
-     1.34±0.01ms          208±1μs     0.16  bench_function_base.Partition.time_argpartition('float32', ('random',), 10)
-     1.35±0.01ms          207±1μs     0.15  bench_function_base.Partition.time_argpartition('float32', ('random',), 1000)
```