@charris 2023-08-02
>  I think unrelated to this patch.

Yes, unrelated. That failure happens now and then, the give away is "and 1 action required checks".